Refactoring onit to implement local cluster
