+++ pkg/onit/cli/get.go
@@ -23,7 +23,8 @@ import (
 	"text/tabwriter"
 	"time"
 
-	"github.com/onosproject/onos-test/pkg/onit"
+	"github.com/onosproject/onos-test/pkg/onit/k8s"
+
 	"github.com/onosproject/onos-test/pkg/runner"
 	"github.com/spf13/cobra"
 	corev1 "k8s.io/api/core/v1"
@@ -118,7 +119,7 @@ func getGetNetworksCommand() *cobra.Command {
 		Short: "Get the currently configured cluster's networks",
 		Run: func(cmd *cobra.Command, args []string) {
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -161,7 +162,7 @@ func getGetSimulatorsCommand() *cobra.Command {
 		Short: "Get the currently configured cluster's simulators",
 		Run: func(cmd *cobra.Command, args []string) {
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -204,7 +205,7 @@ func getGetClustersCommand() *cobra.Command {
 		Short: "Get a list of all deployed clusters",
 		Run: func(cmd *cobra.Command, args []string) {
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -223,7 +224,7 @@ func getGetClustersCommand() *cobra.Command {
 	return cmd
 }
 
-func printClusters(clusters map[string]*onit.ClusterConfig, includeHeaders bool) {
+func printClusters(clusters map[string]*k8s.ClusterConfig, includeHeaders bool) {
 	writer := new(tabwriter.Writer)
 	writer.Init(os.Stdout, 0, 0, 3, ' ', tabwriter.FilterHTML)
 	if includeHeaders {
@@ -268,7 +269,7 @@ func getGetPartitionsCommand() *cobra.Command {
 		Short: "Get a list of partitions in the cluster",
 		Run: func(cmd *cobra.Command, args []string) {
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -301,7 +302,7 @@ func getGetPartitionsCommand() *cobra.Command {
 	return cmd
 }
 
-func printPartitions(partitions []onit.PartitionInfo) {
+func printPartitions(partitions []k8s.PartitionInfo) {
 	writer := new(tabwriter.Writer)
 	writer.Init(os.Stdout, 0, 0, 3, ' ', tabwriter.FilterHTML)
 	fmt.Fprintln(writer, "ID\tGROUP\tNODES")
@@ -318,7 +319,7 @@ func getGetAppsCommand() *cobra.Command {
 		Short: "Get the currently configured cluster's apps",
 		Run: func(cmd *cobra.Command, args []string) {
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -362,7 +363,7 @@ func getGetPartitionCommand() *cobra.Command {
 		Args:  cobra.ExactArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -408,7 +409,7 @@ func getGetNodesCommand() *cobra.Command {
 		Short: "Get a list of nodes in the cluster",
 		Run: func(cmd *cobra.Command, args []string) {
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -431,7 +432,7 @@ func getGetNodesCommand() *cobra.Command {
 			}
 
 			// Get the list of nodes and output
-			if strings.Compare(nodeType, string(onit.OnosAll)) == 0 {
+			if strings.Compare(nodeType, string(k8s.OnosAll)) == 0 {
 				nodes, err := cluster.GetNodes()
 				if err != nil {
 					exitError(err)
@@ -439,7 +440,7 @@ func getGetNodesCommand() *cobra.Command {
 					noHeaders, _ := cmd.Flags().GetBool("no-headers")
 					printNodes(nodes, !noHeaders)
 				}
-			} else if strings.Compare(nodeType, string(onit.OnosConfig)) == 0 {
+			} else if strings.Compare(nodeType, string(k8s.OnosConfig)) == 0 {
 				nodes, err := cluster.GetOnosConfigNodes()
 				if err != nil {
 					exitError(err)
@@ -448,7 +449,7 @@ func getGetNodesCommand() *cobra.Command {
 					printNodes(nodes, !noHeaders)
 				}
 
-			} else if strings.Compare(nodeType, string(onit.OnosTopo)) == 0 {
+			} else if strings.Compare(nodeType, string(k8s.OnosTopo)) == 0 {
 				nodes, err := cluster.GetOnosTopoNodes()
 				if err != nil {
 					exitError(err)
@@ -457,7 +458,7 @@ func getGetNodesCommand() *cobra.Command {
 					printNodes(nodes, !noHeaders)
 				}
 
-			} else if strings.Compare(nodeType, string(onit.OnosCli)) == 0 {
+			} else if strings.Compare(nodeType, string(k8s.OnosCli)) == 0 {
 				nodes, err := cluster.GetOnosCliNodes()
 				if err != nil {
 					exitError(err)
@@ -477,7 +478,7 @@ func getGetNodesCommand() *cobra.Command {
 	return cmd
 }
 
-func printNodes(nodes []onit.NodeInfo, includeHeaders bool) {
+func printNodes(nodes []k8s.NodeInfo, includeHeaders bool) {
 	writer := new(tabwriter.Writer)
 	writer.Init(os.Stdout, 0, 0, 3, ' ', tabwriter.FilterHTML)
 	if includeHeaders {
@@ -581,7 +582,7 @@ func getGetHistoryCommand() *cobra.Command {
 		Short: "Get the history of test runs",
 		Run: func(cmd *cobra.Command, args []string) {
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -616,7 +617,7 @@ func getGetHistoryCommand() *cobra.Command {
 }
 
 // printHistory prints a test history in table format
-func printHistory(records []onit.TestRecord) {
+func printHistory(records []k8s.TestRecord) {
 	writer := new(tabwriter.Writer)
 	writer.Init(os.Stdout, 0, 0, 3, ' ', tabwriter.FilterHTML)
 	fmt.Fprintln(writer, "ID\tTESTS\tSTATUS\tEXIT CODE\tMESSAGE")
@@ -645,7 +646,7 @@ To output the logs from a test, get the test ID from the test run or from 'onit 
 			stream, _ := cmd.Flags().GetBool("stream")
 
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
