+++ pkg/onit/cli/remove.go
@@ -17,7 +17,8 @@ package cli
 import (
 	"errors"
 
-	"github.com/onosproject/onos-test/pkg/onit"
+	"github.com/onosproject/onos-test/pkg/onit/k8s"
+
 	"github.com/spf13/cobra"
 )
 
@@ -56,7 +57,7 @@ func getRemoveNetworkCommand() *cobra.Command {
 			name := args[0]
 
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -105,7 +106,7 @@ func getRemoveSimulatorCommand() *cobra.Command {
 			name := args[0]
 
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
@@ -155,7 +156,7 @@ func getRemoveAppCommand() *cobra.Command {
 			name := args[0]
 
 			// Get the onit controller
-			controller, err := onit.NewController()
+			controller, err := k8s.NewController()
 			if err != nil {
 				exitError(err)
 			}
