+++ pkg/onit/k8s/config.go
@@ -12,23 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package onit
+package k8s
 
 import (
 	"encoding/json"
 	"io/ioutil"
-	"k8s.io/api/core/v1"
 	"os"
 	"path/filepath"
 	"runtime"
+
+	v1 "k8s.io/api/core/v1"
 )
 
 var (
 	_, path, _, _     = runtime.Caller(0)
-	certsPath         = filepath.Join(filepath.Dir(filepath.Dir(path)), "certs")
-	configsPath       = filepath.Join(filepath.Dir(filepath.Dir(path)), "../configs")
-	deviceConfigsPath = filepath.Join(filepath.Join(filepath.Dir(filepath.Dir(path)), "../configs"), "device")
-	storeConfigsPath  = filepath.Join(filepath.Join(filepath.Dir(filepath.Dir(path)), "../configs"), "store")
+	certsPath         = filepath.Join(filepath.Dir(filepath.Dir(path)), "../certs")
+	configsPath       = filepath.Join(filepath.Dir(filepath.Dir(path)), "../../configs")
+	deviceConfigsPath = filepath.Join(filepath.Join(filepath.Dir(filepath.Dir(path)), "../../configs"), "device")
+	storeConfigsPath  = filepath.Join(filepath.Join(filepath.Dir(filepath.Dir(path)), "../../configs"), "store")
 )
 
 // ClusterConfig provides the configuration for the Kubernetes test cluster
