+++ pkg/onit/k8s/onos-config.go
@@ -12,100 +12,143 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package onit
+package k8s
 
 import (
-	"bytes"
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"path/filepath"
-	"strconv"
+	"strings"
 	"time"
 
-	"gopkg.in/yaml.v1"
-
 	"k8s.io/apimachinery/pkg/labels"
-
-	"k8s.io/apimachinery/pkg/util/intstr"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
-// setupOnosTopo sets up the onos-topo Deployment
-func (c *ClusterController) setupOnosTopo() error {
-	if err := c.createOnosTopoConfigMap(); err != nil {
+// setupOnosConfig sets up the onos-config Deployment
+func (c *ClusterController) setupOnosConfig() error {
+	if err := c.createOnosConfigConfigMap(); err != nil {
 		return err
 	}
-	if err := c.createOnosTopoService(); err != nil {
+	if err := c.createOnosConfigService(); err != nil {
 		return err
 	}
-	if err := c.createOnosTopoDeployment(); err != nil {
+	if err := c.createOnosConfigDeployment(); err != nil {
 		return err
 	}
-	if err := c.createOnosTopoProxyConfigMap(); err != nil {
+	if err := c.createOnosConfigProxyConfigMap(); err != nil {
 		return err
 	}
-	if err := c.createOnosTopoProxyDeployment(); err != nil {
+	if err := c.createOnosConfigProxyDeployment(); err != nil {
 		return err
 	}
-	if err := c.createOnosTopoProxyService(); err != nil {
+	if err := c.createOnosConfigProxyService(); err != nil {
 		return err
 	}
-	if err := c.awaitOnosTopoDeploymentReady(); err != nil {
+	if err := c.awaitOnosConfigDeploymentReady(); err != nil {
 		return err
 	}
-	if err := c.awaitOnosTopoProxyDeploymentReady(); err != nil {
+	if err := c.awaitOnosConfigProxyDeploymentReady(); err != nil {
 		return err
 	}
 	return nil
 }
 
-// createOnosTopoConfigMap creates a ConfigMap for the onos-topo Deployment
-func (c *ClusterController) createOnosTopoConfigMap() error {
+// createOnosConfigConfigMap creates a ConfigMap for the onos-config Deployment
+func (c *ClusterController) createOnosConfigConfigMap() error {
+	config, err := c.config.load()
+	if err != nil {
+		return err
+	}
+
+	// Serialize the change store configuration
+	changeStore, err := json.Marshal(config["changeStore"])
+	if err != nil {
+		return err
+	}
+
+	// Serialize the network store configuration
+	networkStore, err := json.Marshal(config["networkStore"])
+	if err != nil {
+		return err
+	}
+
+	// Serialize the device store configuration
+	deviceStore, err := json.Marshal(config["deviceStore"])
+	if err != nil {
+		return err
+	}
+
+	// Serialize the config store configuration
+	configStore, err := json.Marshal(config["configStore"])
+	if err != nil {
+		return err
+	}
+
 	cm := &corev1.ConfigMap{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      "onos-topo",
+			Name:      "onos-config",
 			Namespace: c.clusterID,
 		},
-		Data: map[string]string{},
+		Data: map[string]string{
+			"changeStore.json":  string(changeStore),
+			"configStore.json":  string(configStore),
+			"deviceStore.json":  string(deviceStore),
+			"networkStore.json": string(networkStore),
+		},
 	}
-	_, err := c.kubeclient.CoreV1().ConfigMaps(c.clusterID).Create(cm)
+	_, err = c.kubeclient.CoreV1().ConfigMaps(c.clusterID).Create(cm)
 	return err
 }
 
-// createOnosTopoService creates a Service to expose the onos-topo Deployment to other pods
-func (c *ClusterController) createOnosTopoService() error {
-	service := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "onos-topo",
-			Namespace: c.clusterID,
-		},
-		Spec: corev1.ServiceSpec{
-			Selector: map[string]string{
-				"app":  "onos",
-				"type": "topo",
-			},
-			Ports: []corev1.ServicePort{
-				{
-					Name: "grpc",
-					Port: 5150,
-				},
-			},
-		},
+// createModelPluginString creates model plugin path based on a device type, version, and image tag
+func (c *ClusterController) createModelPluginString(deviceType string, version string, debug bool) string {
+	var sb strings.Builder
+	sb.WriteString("-modelPlugin=/usr/local/lib/")
+	sb.WriteString(deviceType)
+	if debug {
+		sb.WriteString("-debug.so.")
+		sb.WriteString(version)
+	} else {
+		sb.WriteString(".so.")
+		sb.WriteString(version)
 	}
-	_, err := c.kubeclient.CoreV1().Services(c.clusterID).Create(service)
-	return err
+
+	return sb.String()
 }
 
-// createOnosTopoDeployment creates an onos-topo Deployment
-func (c *ClusterController) createOnosTopoDeployment() error {
-	nodes := int32(c.config.TopoNodes)
+// createOnosConfigDeployment creates an onos-config Deployment
+func (c *ClusterController) createOnosConfigDeployment() error {
+	nodes := int32(c.config.ConfigNodes)
 	zero := int64(0)
+
+	testDevModelPluginV1 := ""
+	testDevModelPluginV2 := ""
+	testDevSimModelPluginV1 := ""
+	testStratumModelPluginV2 := ""
+
+	if c.config.ImageTags["config"] == string(Debug) {
+		testDevModelPluginV1 = c.createModelPluginString("testdevice", "1.0.0", true)
+		testDevModelPluginV2 = c.createModelPluginString("testdevice", "2.0.0", true)
+		testDevSimModelPluginV1 = c.createModelPluginString("devicesim", "1.0.0", true)
+		testStratumModelPluginV2 = c.createModelPluginString("stratum", "1.0.0", true)
+
+	} else {
+		testDevModelPluginV1 = c.createModelPluginString("testdevice", "1.0.0", false)
+		testDevModelPluginV2 = c.createModelPluginString("testdevice", "2.0.0", false)
+		testDevSimModelPluginV1 = c.createModelPluginString("devicesim", "1.0.0", false)
+		testStratumModelPluginV2 = c.createModelPluginString("stratum", "1.0.0", false)
+
+	}
+
 	dep := &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      "onos-topo",
+			Name:      "onos-config",
 			Namespace: c.clusterID,
 		},
 		Spec: appsv1.DeploymentSpec{
@@ -113,15 +156,15 @@ func (c *ClusterController) createOnosTopoDeployment() error {
 			Selector: &metav1.LabelSelector{
 				MatchLabels: map[string]string{
 					"app":  "onos",
-					"type": "topo",
+					"type": "config",
 				},
 			},
 			Template: corev1.PodTemplateSpec{
 				ObjectMeta: metav1.ObjectMeta{
 					Labels: map[string]string{
 						"app":      "onos",
-						"type":     "topo",
-						"resource": "onos-topo",
+						"type":     "config",
+						"resource": "onos-config",
 					},
 					Annotations: map[string]string{
 						"seccomp.security.alpha.kubernetes.io/pod": "unconfined",
@@ -130,8 +173,8 @@ func (c *ClusterController) createOnosTopoDeployment() error {
 				Spec: corev1.PodSpec{
 					Containers: []corev1.Container{
 						{
-							Name:            "onos-topo",
-							Image:           c.imageName("onosproject/onos-topo", c.config.ImageTags["topo"]),
+							Name:            "onos-config",
+							Image:           c.imageName("onosproject/onos-config", c.config.ImageTags["config"]),
 							ImagePullPolicy: c.config.PullPolicy,
 							Env: []corev1.EnvVar{
 								{
@@ -152,9 +195,16 @@ func (c *ClusterController) createOnosTopoDeployment() error {
 								},
 							},
 							Args: []string{
-								"-caPath=/etc/onos-topo/certs/onf.cacrt",
-								"-keyPath=/etc/onos-topo/certs/onos-config.key",
-								"-certPath=/etc/onos-topo/certs/onos-config.crt",
+								"-caPath=/etc/onos-config/certs/onf.cacrt",
+								"-keyPath=/etc/onos-config/certs/onos-config.key",
+								"-certPath=/etc/onos-config/certs/onos-config.crt",
+								"-configStore=/etc/onos-config/configs/configStore.json",
+								"-changeStore=/etc/onos-config/configs/changeStore.json",
+								"-networkStore=/etc/onos-config/configs/networkStore.json",
+								testDevModelPluginV1,
+								testDevModelPluginV2,
+								testDevSimModelPluginV1,
+								testStratumModelPluginV2,
 							},
 							Ports: []corev1.ContainerPort{
 								{
@@ -187,13 +237,13 @@ func (c *ClusterController) createOnosTopoDeployment() error {
 							},
 							VolumeMounts: []corev1.VolumeMount{
 								{
-									Name:      "topo",
-									MountPath: "/etc/onos-topo/configs",
+									Name:      "config",
+									MountPath: "/etc/onos-config/configs",
 									ReadOnly:  true,
 								},
 								{
 									Name:      "secret",
-									MountPath: "/etc/onos-topo/certs",
+									MountPath: "/etc/onos-config/certs",
 									ReadOnly:  true,
 								},
 							},
@@ -211,11 +261,11 @@ func (c *ClusterController) createOnosTopoDeployment() error {
 					},
 					Volumes: []corev1.Volume{
 						{
-							Name: "topo",
+							Name: "config",
 							VolumeSource: corev1.VolumeSource{
 								ConfigMap: &corev1.ConfigMapVolumeSource{
 									LocalObjectReference: corev1.LocalObjectReference{
-										Name: "onos-topo",
+										Name: "onos-config",
 									},
 								},
 							},
@@ -237,9 +287,33 @@ func (c *ClusterController) createOnosTopoDeployment() error {
 	return err
 }
 
-// awaitOnosTopoDeploymentReady waits for the onos-topo pods to complete startup
-func (c *ClusterController) awaitOnosTopoDeploymentReady() error {
-	labelSelector := metav1.LabelSelector{MatchLabels: map[string]string{"app": "onos", "type": "topo"}}
+// createOnosConfigService creates a Service to expose the onos-config Deployment to other pods
+func (c *ClusterController) createOnosConfigService() error {
+	service := &corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "onos-config",
+			Namespace: c.clusterID,
+		},
+		Spec: corev1.ServiceSpec{
+			Selector: map[string]string{
+				"app":  "onos",
+				"type": "config",
+			},
+			Ports: []corev1.ServicePort{
+				{
+					Name: "grpc",
+					Port: 5150,
+				},
+			},
+		},
+	}
+	_, err := c.kubeclient.CoreV1().Services(c.clusterID).Create(service)
+	return err
+}
+
+// awaitOnosConfigDeploymentReady waits for the onos-config pods to complete startup
+func (c *ClusterController) awaitOnosConfigDeploymentReady() error {
+	labelSelector := metav1.LabelSelector{MatchLabels: map[string]string{"app": "onos", "type": "config"}}
 	unblocked := make(map[string]bool)
 	for {
 		// Get a list of the pods that match the deployment
@@ -259,50 +333,52 @@ func (c *ClusterController) awaitOnosTopoDeploymentReady() error {
 						return err
 					}
 				}
+
 				unblocked[pod.Name] = true
 			}
 		}
 
-		// Get the onos-topo deployment
-		dep, err := c.kubeclient.AppsV1().Deployments(c.clusterID).Get("onos-topo", metav1.GetOptions{})
+		// Get the onos-config deployment
+		dep, err := c.kubeclient.AppsV1().Deployments(c.clusterID).Get("onos-config", metav1.GetOptions{})
 		if err != nil {
 			return err
 		}
 
 		// Return once the all replicas in the deployment are ready
-		if int(dep.Status.ReadyReplicas) == c.config.TopoNodes {
+		if int(dep.Status.ReadyReplicas) == c.config.ConfigNodes {
 			return nil
 		}
 		time.Sleep(100 * time.Millisecond)
 	}
 }
 
-// createOnosTopoProxyConfigMap creates a ConfigMap for the onos-topo-envoy Deployment
-func (c *ClusterController) createOnosTopoProxyConfigMap() error {
-	configPath := filepath.Join(filepath.Join(configsPath, "envoy"), "envoy-topo.yaml")
+// createOnosConfigProxyConfigMap creates a ConfigMap for the onos-config-envoy Deployment
+func (c *ClusterController) createOnosConfigProxyConfigMap() error {
+	configPath := filepath.Join(filepath.Join(configsPath, "envoy"), "envoy-config.yaml")
 	data, err := ioutil.ReadFile(configPath)
 	if err != nil {
 		return err
 	}
+
 	cm := &corev1.ConfigMap{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      "onos-topo-envoy",
+			Name:      "onos-config-envoy",
 			Namespace: c.clusterID,
 		},
 		BinaryData: map[string][]byte{
-			"envoy-topo.yaml": data,
+			"envoy-config.yaml": data,
 		},
 	}
 	_, err = c.kubeclient.CoreV1().ConfigMaps(c.clusterID).Create(cm)
 	return err
 }
 
-// createOnosTopoProxyDeployment creates an onos-topo Envoy proxy
-func (c *ClusterController) createOnosTopoProxyDeployment() error {
+// createOnosConfigProxyDeployment creates an onos-config Envoy proxy
+func (c *ClusterController) createOnosConfigProxyDeployment() error {
 	nodes := int32(1)
 	deployment := &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      "onos-topo-envoy",
+			Name:      "onos-config-envoy",
 			Namespace: c.clusterID,
 		},
 		Spec: appsv1.DeploymentSpec{
@@ -310,27 +386,27 @@ func (c *ClusterController) createOnosTopoProxyDeployment() error {
 			Selector: &metav1.LabelSelector{
 				MatchLabels: map[string]string{
 					"app":  "onos",
-					"type": "topo-envoy",
+					"type": "config-envoy",
 				},
 			},
 			Template: corev1.PodTemplateSpec{
 				ObjectMeta: metav1.ObjectMeta{
 					Labels: map[string]string{
 						"app":      "onos",
-						"type":     "topo-envoy",
-						"resource": "onos-topo",
+						"type":     "config-envoy",
+						"resource": "onos-config",
 					},
 				},
 				Spec: corev1.PodSpec{
 					Containers: []corev1.Container{
 						{
-							Name:            "onos-topo-envoy",
+							Name:            "onos-config-envoy",
 							Image:           "envoyproxy/envoy-alpine:latest",
 							ImagePullPolicy: c.config.PullPolicy,
 							Command: []string{
 								"/usr/local/bin/envoy",
 								"-c",
-								"/etc/envoy-proxy/config/envoy-topo.yaml",
+								"/etc/envoy-proxy/config/envoy-config.yaml",
 							},
 							Ports: []corev1.ContainerPort{
 								{
@@ -358,7 +434,7 @@ func (c *ClusterController) createOnosTopoProxyDeployment() error {
 							VolumeSource: corev1.VolumeSource{
 								ConfigMap: &corev1.ConfigMapVolumeSource{
 									LocalObjectReference: corev1.LocalObjectReference{
-										Name: "onos-topo-envoy",
+										Name: "onos-config-envoy",
 									},
 								},
 							},
@@ -380,17 +456,17 @@ func (c *ClusterController) createOnosTopoProxyDeployment() error {
 	return err
 }
 
-// createOnosTopoProxyService creates an onos-topo Envoy proxy service
-func (c *ClusterController) createOnosTopoProxyService() error {
+// createOnosConfigProxyService creates an onos-config Envoy proxy service
+func (c *ClusterController) createOnosConfigProxyService() error {
 	service := &corev1.Service{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      "onos-topo-envoy",
+			Name:      "onos-config-envoy",
 			Namespace: c.clusterID,
 		},
 		Spec: corev1.ServiceSpec{
 			Selector: map[string]string{
 				"app":  "onos",
-				"type": "topo-envoy",
+				"type": "config-envoy",
 			},
 			Ports: []corev1.ServicePort{
 				{
@@ -404,11 +480,11 @@ func (c *ClusterController) createOnosTopoProxyService() error {
 	return err
 }
 
-// awaitOnosTopoProxyDeploymentReady waits for the onos-topo proxy pods to complete startup
-func (c *ClusterController) awaitOnosTopoProxyDeploymentReady() error {
+// awaitOnosConfigProxyDeploymentReady waits for the onos-config proxy pods to complete startup
+func (c *ClusterController) awaitOnosConfigProxyDeploymentReady() error {
 	for {
-		// Get the onos-topo deployment
-		dep, err := c.kubeclient.AppsV1().Deployments(c.clusterID).Get("onos-topo-envoy", metav1.GetOptions{})
+		// Get the onos-config deployment
+		dep, err := c.kubeclient.AppsV1().Deployments(c.clusterID).Get("onos-config-envoy", metav1.GetOptions{})
 		if err != nil {
 			return err
 		}
@@ -421,80 +497,18 @@ func (c *ClusterController) awaitOnosTopoProxyDeploymentReady() error {
 	}
 }
 
-// addSimulatorToTopo adds a simulator to onos-topo
-func (c *ClusterController) addSimulatorToTopo(name string) error {
-	return c.addDevice("Devicesim", name, 11161)
-}
-
-// addNetworkToTopo adds a network to onos-topo
-func (c *ClusterController) addNetworkToTopo(name string, config *NetworkConfig) error {
-	var port = 50001
-	for i := 0; i < config.NumDevices; i++ {
-		var buf bytes.Buffer
-		buf.WriteString(name)
-		buf.WriteString("-s")
-		buf.WriteString(strconv.Itoa(i))
-		deviceName := buf.String()
-		if err := c.addDevice("Stratum", deviceName, port); err != nil {
-			return err
-		}
-		port = port + 1
-	}
-	return nil
-}
-
-// addDevice adds the given device via the CLI
-func (c *ClusterController) addDevice(deviceType string, name string, port int) error {
-	command := fmt.Sprintf("onos topo add device %s --type %s --address %s:%d --version 1.0.0 --plain --timeout 15s", name, deviceType, name, port)
-	return c.executeCLI(command)
-}
-
-// removeSimulatorFromConfig removes a simulator from the onos-config configuration
-func (c *ClusterController) removeSimulatorFromConfig(name string) error {
-	return c.removeDevice(name)
-}
-
-// removeNetworkFromConfig removes a network from the onos-config configuration
-func (c *ClusterController) removeNetworkFromConfig(name string, configMap *corev1.ConfigMapList) error {
-	dataMap := configMap.Items[0].BinaryData["config"]
-	m := make(map[string]interface{})
-	err := yaml.Unmarshal(dataMap, &m)
-	if err != nil {
-		return err
-	}
-	numDevices := m["numdevices"].(int)
-
-	for i := 0; i < numDevices; i++ {
-		var buf bytes.Buffer
-		buf.WriteString(name)
-		buf.WriteString("-s")
-		buf.WriteString(strconv.Itoa(i))
-		deviceName := buf.String()
-		if err = c.removeDevice(deviceName); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// removeDevice removes the given device via the given pod
-func (c *ClusterController) removeDevice(name string) error {
-	command := fmt.Sprintf("onos topo remove device %s", name)
-	return c.executeCLI(command)
-}
-
-// GetOnosTopoNodes returns a list of all onos-topo nodes running in the cluster
-func (c *ClusterController) GetOnosTopoNodes() ([]NodeInfo, error) {
-	topoLabelSelector := metav1.LabelSelector{MatchLabels: map[string]string{"app": "onos", "type": "topo"}}
+// GetOnosConfigNodes returns a list of all onos-config nodes running in the cluster
+func (c *ClusterController) GetOnosConfigNodes() ([]NodeInfo, error) {
+	configLabelSelector := metav1.LabelSelector{MatchLabels: map[string]string{"app": "onos", "type": "config"}}
 
 	pods, err := c.kubeclient.CoreV1().Pods(c.clusterID).List(metav1.ListOptions{
-		LabelSelector: labels.Set(topoLabelSelector.MatchLabels).String(),
+		LabelSelector: labels.Set(configLabelSelector.MatchLabels).String(),
 	})
 	if err != nil {
 		return nil, err
 	}
 
-	onosTopoNodes := make([]NodeInfo, len(pods.Items))
+	onosConfigNodes := make([]NodeInfo, len(pods.Items))
 	for i, pod := range pods.Items {
 		var status NodeStatus
 		if pod.Status.Phase == corev1.PodRunning {
@@ -502,12 +516,12 @@ func (c *ClusterController) GetOnosTopoNodes() ([]NodeInfo, error) {
 		} else if pod.Status.Phase == corev1.PodFailed {
 			status = NodeFailed
 		}
-		onosTopoNodes[i] = NodeInfo{
+		onosConfigNodes[i] = NodeInfo{
 			ID:     pod.Name,
 			Status: status,
-			Type:   OnosTopo,
+			Type:   OnosConfig,
 		}
 	}
 
-	return onosTopoNodes, nil
+	return onosConfigNodes, nil
 }
